+++ srcs/front/vitejs/src/classes/Friend.tsx
@@ -3,10 +3,13 @@ import User from "./User";
 import Message from "./Message";
 import FriendRequest from "./FriendRequest";
 
+const MESSAGE_RECIVED = 20;
+
 class Friend extends User {
     messages: Message[];
     connected: boolean;
     nbNotifs: number;
+
 
 
     constructor(id: number, name: string, email: string, profPicture: string, bio: string, lang : string,
@@ -86,7 +89,7 @@ class Friend extends User {
         }
     }
 
-    static async getFriends() : Promise<Friend[] | undefined> {
+    static async getFriends() : Promise<Friend[] | undefined> { //socket: WebSocket | null
 
         try {
             const requestData : RequestInit = {
@@ -106,10 +109,38 @@ class Friend extends User {
 
             for (const req of dataReponse) {
                 const user : User | undefined = await Friend.getUserById(req.friendUserId);
+
+                // action: string;
+                // targetId: number;
+                // skip?: number;
+                // take?: number;
+                // msg?: string;
+
+                // chercher ici tout les messages
+                // console.log("debut de la recuperation de messge...");
+                // if (user && socket && socket.readyState === WebSocket.OPEN) {
+                //     try {
+                //         let i = 0;
+                //         let buffer = [];
+                //         // console.log(user);
+                //         do {
+                //             console.log("recuperation de messge...");
+                            
+                            
+                //             socket.send(JSON.stringify({ action: 'refresh', targetId: user.id, take: MESSAGE_RECIVED, skip: i * MESSAGE_RECIVED}));
+                //             i++;
+                //         } while (condition);
+            
+                //         // return (new Message(idSender, targetId, new Date(), message));
+                //     } catch (error) {
+                //         console.error("Erreur lors de l'envoi de la demande des requetes :", error);
+                //     }
+                // }
                 
                 if (user != undefined)
                     friends.push(new Friend(req.friendUserId, user.name, user.email, user.profPicture, user.bio, user.lang, user.isTwoFactorEnabled, user.rank));
             }
+            
 
             return (friends);
